Refactor hashing behavior into fshash.
This behavior too will be more broadly reusable than just the 'dir' input implementation.

Clean up some other file organization while we're there (MemoryBucket implementation has gotten big enough to deserve its own file).
