+++ input/dir/dir_input.go
@@ -11,7 +11,6 @@ import (
 	"polydawn.net/repeatr/def"
 	"polydawn.net/repeatr/input"
 	"polydawn.net/repeatr/lib/fshash"
-	"polydawn.net/repeatr/lib/treewalk"
 )
 
 const Type = "dir"
@@ -52,26 +51,9 @@ func (i Input) Apply(destinationRoot string) <-chan error {
 		}
 
 		// hash whole tree
-		hasher := i.hasherFactory()
-		preVisit := func(node treewalk.Node) error {
-			record := node.(fshash.RecordIterator).Record()
-			metabin, err := record.Metadata.MarshalBinary()
-			if err != nil {
-				return err
-			}
-			// TODO : this also needs some higher level alignment/length stuff
-			_, err = hasher.Write(metabin)
-			if err != nil {
-				return err
-			}
-			_, err = hasher.Write(record.ContentHash)
-			return err
-		}
-		postVisit := func(node treewalk.Node) error { return nil }
-		err = treewalk.Walk(bucket.Iterator(), preVisit, postVisit)
+		actualTreeHash, _ := fshash.Hash(bucket, i.hasherFactory())
 
 		// verify total integrity
-		actualTreeHash := hasher.Sum(nil)
 		expectedTreeHash, err := base64.URLEncoding.DecodeString(i.spec.Hash)
 		if !bytes.Equal(actualTreeHash, expectedTreeHash) {
 			done <- input.InputHashMismatchError.New("expected hash %q, got %q", i.spec.Hash, base64.URLEncoding.EncodeToString(actualTreeHash))
